feat(scripts): use cycletimer + move it to utils
