+++ docs/source/il_robots.mdx
@@ -412,6 +412,29 @@ Avoid adding too much variation too quickly, as it may hinder your results.
 
 If you want to dive deeper into this important topic, you can check out the [blog post](https://huggingface.co/blog/lerobot-datasets#what-makes-a-good-dataset) we wrote on what makes a good dataset.
 
+#### Cadence reporting
+
+`lerobot-record` paces the loop at `--dataset.fps` and reports what it actually achieved: a one-line digest for every episode you keep, and a block for the whole session when recording ends — including on `Ctrl-C`.
+
+```
+Cadence (episode 0): 29.88 Hz vs 30 Hz target · 600 ticks, 20.0 s measured · 8/599 ticks over the 33.3 ms budget (work mean 18.4 ms, worst 45.1 ms)
+Cadence (episode 1): 29.88 Hz vs 30 Hz target · 600 ticks, 20.0 s measured · 8/599 ticks over the 33.3 ms budget (work mean 18.4 ms, worst 45.1 ms)
+Cadence summary — whole run, 2 episodes · target 30 Hz (33.3 ms budget per tick): 1200 ticks, 1198 judged
+  effective cadence: 29.88 Hz over 40.1 s measured
+  ticks over the 33.3 ms work budget: 16/1198 (1.3%) — work mean 18.4 ms, worst 45.1 ms
+  loop-body steps (share of measured work):
+    observe      mean  13.65 ms · worst  40.20 ms ·  74.4% of work · 1200 calls
+    process_obs  mean   0.30 ms · worst   0.30 ms ·   1.6% of work · 1200 calls
+    teleop       mean   1.80 ms · worst   1.80 ms ·   9.8% of work · 1200 calls
+    send         mean   0.90 ms · worst   0.90 ms ·   4.9% of work · 1200 calls
+    record       mean   1.70 ms · worst   1.90 ms ·   9.3% of work · 1200 calls
+  pacing headroom: 15.1 ms slept per tick on average (max 16.7 ms) — near zero means the loop is saturated
+```
+
+This is worth reading before you train on the dataset. A frame's timestamp is derived from its index, so a session that ran at 25 Hz still produces a dataset that _claims_ 30 — the recorded motion is simply faster than reality. Above, `observe` is 74% of the loop's work and its `worst` is 40 ms on a 33.3 ms budget: the cameras are what pushed those 16 ticks over. **Pacing headroom** near zero means there is nothing left to absorb a slow tick.
+
+The reset phases between episodes are paced the same way but are deliberately not part of these numbers, since they record nothing. `lerobot-teleoperate` and `lerobot-replay` report the same way, and `lerobot-rollout` adds the interpolation-specific lines described in [Cadence Reporting](./inference#cadence-reporting).
+
 #### Troubleshooting:
 
 - On Linux, the recording control-flow keys (arrow keys, Escape) work on X11, Wayland, and headless/SSH sessions as long as `lerobot-record` runs in an interactive terminal — no `$DISPLAY` setup is needed. If the keys have no effect, make sure you are in an interactive (TTY) terminal, not a piped/non-TTY session, and that it is focused; the letter equivalents `n` / `r` / `q` also work. Keyboard _teleoperation_ (as opposed to the recording control flow) still requires a global key backend — an X11 session, a Windows desktop, or macOS with Accessibility/Input Monitoring granted — and is unavailable on Wayland or headless machines. See [pynput limitations](https://pynput.readthedocs.io/en/latest/limitations.html#linux).
